refactor: remove limit for now and cleanup
