--- components/LocalTime/index.ts
@@ -1,2 +1,0 @@
-export * from "./LocalTime";
-export * from "./utils";
